+++ fe/fe-common/src/main/java/org/apache/doris/common/credentials/DataLakeAWSCredentialsProvider.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.datasource.credentials;
+package org.apache.doris.common.credentials;
 
 import com.amazonaws.SdkClientException;
 import com.amazonaws.auth.AWSCredentials;
