+++ fe/fe-core/src/main/java/org/apache/doris/common/util/S3Util.java
@@ -17,7 +17,7 @@
 
 package org.apache.doris.common.util;
 
-import org.apache.doris.datasource.credentials.CloudCredential;
+import org.apache.doris.common.credentials.CloudCredential;
 
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentials;
