+++ fe/fe-core/src/main/java/org/apache/doris/datasource/property/constants/GCSProperties.java
@@ -17,7 +17,7 @@
 
 package org.apache.doris.datasource.property.constants;
 
-import org.apache.doris.datasource.credentials.CloudCredential;
+import org.apache.doris.common.credentials.CloudCredential;
 
 import java.util.Arrays;
 import java.util.List;
